[Silver III] Title: 팩토리얼 나누기, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Silver/15996. 팩토리얼 나누기/팩토리얼 나누기.cc
@@ -2,12 +2,12 @@
 #include <iostream>
 using namespace std;
 
-long n, a, ans;
+long long n, a, ans;
 
 int main(void) {
     ios::sync_with_stdio(false), cin.tie(NULL);
     cin >> n >> a;
-    for (long i = a; i <= n; i *= a) {
+    for (long long i = a; i <= n; i *= a) {
         ans += n / i;
     }
     cout << ans;
